Delete blank-java and move its contents to java-events.
- The blank-java example is just a Lambda function that handles SQSEvent objects.
- Moved this example to inside the java-events folder.
- The Lambda developer guide now must remove all references to blank-java.
- Tested this change using the sqs path, as well as ensured current examples still work.
