--- sample-apps/blank-java/2-build-layer.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-set -eo pipefail
-gradle -q packageLibs
-mv build/distributions/blank-java.zip build/blank-java-lib.zip
