--- sample-apps/blank-java/4-invoke.sh
@@ -1,10 +1,0 @@
-#!/bin/bash
-set -eo pipefail
-FUNCTION=$(aws cloudformation describe-stack-resource --stack-name blank-java --logical-resource-id function --query 'StackResourceDetail.PhysicalResourceId' --output text)
-
-while true; do
-  aws lambda invoke --function-name $FUNCTION --payload file://event.json out.json
-  cat out.json
-  echo ""
-  sleep 2
-done
