--- sample-apps/blank-java/README.md
@@ -1,96 +1,0 @@
-# Blank function (Java)
-
-![Architecture](/sample-apps/blank-java/images/sample-blank-java.png)
-
-The project source includes function code and supporting resources:
-
-- `src/main` - A Java function.
-- `src/test` - A unit test and helper classes.
-- `template.yml` - An AWS CloudFormation template that creates an application.
-- `build.gradle` - A Gradle build file.
-- `pom.xml` - A Maven build file.
-- `1-create-bucket.sh`, `2-build-layer.sh`, etc. - Shell scripts that use the AWS CLI to deploy and manage the application.
-
-Use the following instructions to deploy the sample application.
-
-# Requirements
-- [Java 8 runtime environment (SE JRE)](https://www.oracle.com/java/technologies/javase-downloads.html)
-- [Gradle 5](https://gradle.org/releases/) or [Maven 3](https://maven.apache.org/docs/history.html)
-- The Bash shell. For Linux and macOS, this is included by default. In Windows 10, you can install the [Windows Subsystem for Linux](https://docs.microsoft.com/en-us/windows/wsl/install-win10) to get a Windows-integrated version of Ubuntu and Bash.
-- [The AWS CLI](https://docs.aws.amazon.com/cli/latest/userguide/cli-chap-install.html) v1.17 or newer.
-
-If you use the AWS CLI v2, add the following to your [configuration file](https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-files.html) (`~/.aws/config`):
-
-```
-cli_binary_format=raw-in-base64-out
-```
-
-This setting enables the AWS CLI v2 to load JSON events from a file, matching the v1 behavior.
-
-# Setup
-Download or clone this repository.
-
-    $ git clone https://github.com/awsdocs/aws-lambda-developer-guide.git
-    $ cd aws-lambda-developer-guide/sample-apps/blank-java
-
-To create a new bucket for deployment artifacts, run `1-create-bucket.sh`.
-
-    blank-java$ ./1-create-bucket.sh
-    make_bucket: lambda-artifacts-a5e491dbb5b22e0d
-
-To build a Lambda layer that contains the function's runtime dependencies, run `2-build-layer.sh`. Packaging dependencies in a layer reduces the size of the deployment package that you upload when you modify your code.
-
-    blank-java$ ./2-build-layer.sh
-
-# Deploy
-
-To deploy the application, run `3-deploy.sh`.
-
-    blank-java$ ./3-deploy.sh
-    BUILD SUCCESSFUL in 1s
-    Successfully packaged artifacts and wrote output template to file out.yml.
-    Waiting for changeset to be created..
-    Successfully created/updated stack - blank-java
-
-This script uses AWS CloudFormation to deploy the Lambda functions and an IAM role. If the AWS CloudFormation stack that contains the resources already exists, the script updates it with any changes to the template or function code.
-
-You can also build the application with Maven. To use maven, add `mvn` to the command.
-
-    java-basic$ ./3-deploy.sh mvn
-    [INFO] Scanning for projects...
-    [INFO] -----------------------< com.example:blank-java >-----------------------
-    [INFO] Building blank-java-function 1.0-SNAPSHOT
-    [INFO] --------------------------------[ jar ]---------------------------------
-    ...
-
-# Test
-To invoke the function, run `4-invoke.sh`.
-
-    blank-java$ ./4-invoke.sh
-    {
-        "StatusCode": 200,
-        "ExecutedVersion": "$LATEST"
-    }
-
-Let the script invoke the function a few times and then press `CRTL+C` to exit.
-
-The application uses AWS X-Ray to trace requests. Open the [X-Ray console](https://console.aws.amazon.com/xray/home#/service-map) to view the service map.
-
-![Service Map](/sample-apps/blank-java/images/blank-java-servicemap.png)
-
-Choose a node in the main function graph. Then choose **View traces** to see a list of traces. Choose any trace to view a timeline that breaks down the work done by the function.
-
-![Trace](/sample-apps/blank-java/images/blank-java-trace.png)
-
-Finally, view the application in the Lambda console.
-
-*To view the application*
-1. Open the [applications page](https://console.aws.amazon.com/lambda/home#/applications) in the Lambda console.
-2. Choose **blank-java**.
-
-  ![Application](/sample-apps/blank-java/images/blank-java-application.png)
-
-# Cleanup
-To delete the application, run `5-cleanup.sh`.
-
-    blank$ ./5-cleanup.sh
