--- sample-apps/blank-java/src/main/resources/log4j2.xml
@@ -1,16 +1,0 @@
-<Configuration status="WARN">
-  <Appenders>
-    <Lambda name="Lambda">
-      <PatternLayout>
-          <pattern>%d{yyyy-MM-dd HH:mm:ss} %X{AWSRequestId} %-5p %c{1} - %m%n</pattern>
-      </PatternLayout>
-    </Lambda>
-  </Appenders>
-  <Loggers>
-    <Root level="INFO">
-      <AppenderRef ref="Lambda"/>
-    </Root>
-    <Logger name="software.amazon.awssdk" level="WARN" />
-    <Logger name="software.amazon.awssdk.request" level="DEBUG" />
-  </Loggers>
-</Configuration>
