--- sample-apps/blank-java/src/test/resources/log4j2.xml
@@ -1,14 +1,0 @@
-<Configuration status="WARN">
-  <Appenders>
-    <Console name="ConsoleAppender" target="SYSTEM_OUT">
-      <PatternLayout pattern="%d{YYYY-MM-dd HH:mm:ss} [%t] %-5p %c:%L - %m%n" />
-    </Console>
-  </Appenders>
-  <Loggers>
-    <Root level="INFO">
-      <AppenderRef ref="ConsoleAppender"/>
-    </Root>
-    <Logger name="software.amazon.awssdk" level="WARN" />
-    <Logger name="software.amazon.awssdk.request" level="DEBUG" />
-  </Loggers>
-</Configuration>
