--- sample-apps/blank-java/template-mvn.yml
@@ -1,20 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: 'AWS::Serverless-2016-10-31'
-Description: An AWS Lambda application that calls the Lambda API.
-Resources:
-  function:
-    Type: AWS::Serverless::Function
-    Properties:
-      CodeUri: target/blank-java-1.0-SNAPSHOT.jar
-      Handler: example.Handler::handleRequest
-      Runtime: java8
-      Description: Java function
-      MemorySize: 512
-      Timeout: 10
-      # Function's execution role
-      Policies:
-        - AWSLambdaBasicExecutionRole
-        - AWSLambda_ReadOnlyAccess
-        - AWSXrayWriteOnlyAccess
-        - AWSLambdaVPCAccessExecutionRole
-      Tracing: Active
