--- sample-apps/blank-java/template.yml
@@ -1,30 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: 'AWS::Serverless-2016-10-31'
-Description: An AWS Lambda application that calls the Lambda API.
-Resources:
-  function:
-    Type: AWS::Serverless::Function
-    Properties:
-      CodeUri: build/distributions/blank-java.zip
-      Handler: example.Handler::handleRequest
-      Runtime: java8
-      Description: Java function
-      MemorySize: 512
-      Timeout: 10
-      # Function's execution role
-      Policies:
-        - AWSLambdaBasicExecutionRole
-        - AWSLambda_ReadOnlyAccess
-        - AWSXrayWriteOnlyAccess
-        - AWSLambdaVPCAccessExecutionRole
-      Tracing: Active
-      Layers:
-        - !Ref libs
-  libs:
-    Type: AWS::Serverless::LayerVersion
-    Properties:
-      LayerName: blank-java-lib
-      Description: Dependencies for the blank-java sample app.
-      ContentUri: build/blank-java-lib.zip
-      CompatibleRuntimes:
-        - java8
