+++ sample-apps/java-events/build.gradle
@@ -8,7 +8,7 @@ repositories {
 
 dependencies {
     implementation 'com.amazonaws:aws-lambda-java-core:1.2.1'
-    implementation 'com.amazonaws:aws-lambda-java-events:3.9.0'
+    implementation 'com.amazonaws:aws-lambda-java-events:3.11.0'
     implementation 'com.google.code.gson:gson:2.8.6'
     implementation 'org.apache.logging.log4j:log4j-api:[2.17.1,)'
     implementation 'org.apache.logging.log4j:log4j-core:[2.17.1,)'
