+++ sample-apps/java-events/pom.xml
@@ -20,7 +20,7 @@
     <dependency>
       <groupId>com.amazonaws</groupId>
       <artifactId>aws-lambda-java-events</artifactId>
-      <version>3.9.0</version>
+      <version>3.11.0</version>
     </dependency>
     <dependency>
       <groupId>com.google.code.gson</groupId>
